+++ packages/blade/src/components/Table/TableHeader.web.tsx
@@ -89,7 +89,8 @@ const StyledHeaderCell = styled(HeaderCell)<{
   $isSortable: boolean;
   $backgroundColor: TableBackgroundColors;
   $rowDensity: NonNullable<TableProps<unknown>['rowDensity']>;
-}>(({ theme, $isSortable, $backgroundColor, $rowDensity }) => ({
+  $hasPadding: boolean;
+}>(({ theme, $isSortable, $backgroundColor, $rowDensity, $hasPadding }) => ({
   '&&&': {
     height: '100%',
     backgroundColor: getIn(theme.colors, $backgroundColor),
@@ -107,15 +108,23 @@ const StyledHeaderCell = styled(HeaderCell)<{
       justifyContent: 'space-between',
       alignItems: 'center',
       height: '100%',
-      paddingLeft: makeSpace(getIn(theme, tableRow.paddingLeft[$rowDensity])),
-      paddingRight: makeSpace(getIn(theme, tableRow.paddingRight[$rowDensity])),
+      paddingLeft: $hasPadding
+        ? makeSpace(getIn(theme, tableRow.paddingLeft[$rowDensity]))
+        : undefined,
+      paddingRight: $hasPadding
+        ? makeSpace(getIn(theme, tableRow.paddingRight[$rowDensity]))
+        : undefined,
       minHeight: makeSize(getIn(size, tableRow.minHeight[$rowDensity])),
     },
     '&:focus-visible': getFocusRingStyles({ theme, negativeOffset: true }),
   },
 }));
 
-const _TableHeaderCell = ({ children, headerKey }: TableHeaderCellProps): React.ReactElement => {
+const _TableHeaderCell = ({
+  children,
+  headerKey,
+  _hasPadding = true,
+}: TableHeaderCellProps): React.ReactElement => {
   const {
     toggleSort,
     currentSortedState,
@@ -132,6 +141,7 @@ const _TableHeaderCell = ({ children, headerKey }: TableHeaderCellProps): React.
       $isSortable={isSortable}
       $backgroundColor={backgroundColor}
       $rowDensity={headerRowDensity ?? rowDensity}
+      $hasPadding={_hasPadding}
       onClick={() => {
         if (isSortable) {
           toggleSort(headerKey);
@@ -164,17 +174,24 @@ const TableHeaderCell = assignWithoutSideEffects(_TableHeaderCell, {
 
 const TableHeaderCellCheckbox = ({
   isChecked,
+  isDisabled,
   isIndeterminate,
   onChange,
 }: {
   isChecked: CheckboxProps['isChecked'];
+  isDisabled: CheckboxProps['isDisabled'];
   isIndeterminate?: CheckboxProps['isIndeterminate'];
   onChange: CheckboxProps['onChange'];
 }): React.ReactElement => {
   return (
     <TableHeaderCell headerKey="SELECT">
       <BaseBox display="flex" alignItems="center" justifyContent="center" flex={1}>
-        <Checkbox isChecked={isChecked} isIndeterminate={isIndeterminate} onChange={onChange} />
+        <Checkbox
+          isChecked={isChecked}
+          isDisabled={isDisabled}
+          isIndeterminate={isIndeterminate}
+          onChange={onChange}
+        />
       </BaseBox>
     </TableHeaderCell>
   );
@@ -197,16 +214,19 @@ const StyledHeaderRow = styled(HeaderRow)<{ $showBorderedCells: boolean }>(
 
 const _TableHeaderRow = ({ children, rowDensity }: TableHeaderRowProps): React.ReactElement => {
   const {
+    disabledRows,
     selectionType,
     selectedRows,
     totalItems,
     toggleAllRowsSelection,
     setHeaderRowDensity,
     showBorderedCells,
+    hasHoverActions,
   } = useTableContext();
   const isMultiSelect = selectionType === 'multiple';
   const isAllSelected = selectedRows && selectedRows.length === totalItems;
   const isIndeterminate = selectedRows && selectedRows.length > 0 && !isAllSelected;
+  const isDisabled = disabledRows && disabledRows.length === totalItems;
   if (rowDensity) {
     setHeaderRowDensity(rowDensity);
   }
@@ -219,11 +239,13 @@ const _TableHeaderRow = ({ children, rowDensity }: TableHeaderRowProps): React.R
       {isMultiSelect && (
         <TableHeaderCellCheckbox
           isChecked={isAllSelected}
+          isDisabled={isDisabled}
           isIndeterminate={isIndeterminate}
           onChange={() => toggleAllRowsSelection()}
         />
       )}
       {children}
+      {hasHoverActions ? <TableHeaderCell _hasPadding={false}>Actions</TableHeaderCell> : null}
     </StyledHeaderRow>
   );
 };